requirements.txt
openai-whisper
sounddevice
numpy
pyttsx3
ultralytics
opencv-python
requests